factor out the wizard buttons to a component
